move tests to folder Issue380
