+++ tests/ORM/Functional/Driver/Common/Integration/Issue380/CaseTest.php
