+++ tests/ORM/Functional/Driver/Common/Integration/Issue380/Entity/User/Email.php
