+++ tests/ORM/Functional/Driver/Common/Integration/Issue380/schema.php
