+++ tests/ORM/Functional/Driver/MySQL/Integration/Issue380/CaseTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Cycle\ORM\Tests\Functional\Driver\MySQL\Integration\Case4;
+namespace Cycle\ORM\Tests\Functional\Driver\MySQL\Integration\Issue380;
 
 // phpcs:ignore
 use Cycle\ORM\Tests\Functional\Driver\Common\Integration\Case4\CaseTest as CommonClass;
