+++ tests/ORM/Functional/Driver/SQLServer/Integration/Issue380/CaseTest.php
@@ -2,16 +2,16 @@
 
 declare(strict_types=1);
 
-namespace Cycle\ORM\Tests\Functional\Driver\Postgres\Integration\Case4;
+namespace Cycle\ORM\Tests\Functional\Driver\SQLServer\Integration\Issue380;
 
 // phpcs:ignore
 use Cycle\ORM\Tests\Functional\Driver\Common\Integration\Case4\CaseTest as CommonClass;
 
 /**
  * @group driver
- * @group driver-postgres
+ * @group driver-sqlserver
  */
 class CaseTest extends CommonClass
 {
-    public const DRIVER = 'postgres';
+    public const DRIVER = 'sqlserver';
 }
